refactor: FeatureFlagProvider
